pyproject.toml
[tool.poetry]
name = "chapter-sync"
version = "0.1.0"
description = ""
authors = [
    "Dan Cardin <ddcardin@gmail.com>",
]
license = "Apache-2.0"
repository = "https://github.com/dancardin/sqlalchemy-declarative-extensions"
readme = 'README.md'

keywords = [
]

include = [
    "py.typed",
    "*.md",
]
packages = [
    { include = "chapter_sync", from = "src" },
]

[tool.poetry.scripts]
chapter-sync = "chapter_sync.__main__:run"

[tool.poetry.dependencies]
python = "^3.10"

Pillow = "^9.0.0"
beautifulsoup4 = "^4.9.3"
cappa = ">=0.16.6"
html5lib = "^1.1"
pendulum = ">=2"
python-dotenv = "*"
requests = "*"
rich = "*"
sqlalchemy = ">=2.0"
xmltodict = "*"

[tool.poetry.group.dev.dependencies]
coverage = "^6.0"
mypy = "1.8.0"
pytest = "^7.2.2"
responses = "^0.23.1"
ruff = "^0.2.2"
sqlalchemy-model-factory = "*"
time-machine = "~2.13.0"
types-beautifulsoup4 = "^4.12.0.20240229"
types-pillow = "^10.2.0.20240213"
types-requests = "^2.31.0.20240218"
types-xmltodict = "^0.13.0.3"

[tool.ruff]
target-version = "py310"

[tool.ruff.lint.isort]
known-first-party = ["chapter_sync", "tests"]

[tool.ruff.lint]
select = ["C", "D", "E", "F", "I", "N", "Q", "RET", "RUF", "S", "T", "UP", "YTT"]
ignore = [
    "C901",
    "E501",
    "S101",
    "D1",
    "D203",
    "D213",
    "D406",
    "D407",
    "D408",
    "D409",
    "D413",
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["T201"]
"src/cappa/parser.py" = ["N818"]

[tool.ruff.lint.pyupgrade]
keep-runtime-typing = true

[tool.coverage.report]
show_missing = true
skip_covered = true
exclude_lines = [
  "pragma: no cover",
  "if TYPE_CHECKING:",
  "if __name__ == .__main__.:",
]

[tool.coverage.run]
include = ["src/*"]
branch = true

[tool.pytest.ini_options]
doctest_optionflags = "NORMALIZE_WHITESPACE IGNORE_EXCEPTION_DETAIL ELLIPSIS"
addopts = "--doctest-modules -vv --ff --strict-markers"
norecursedirs = ".* build dist *.egg bin --junitxml=junit.xml"
filterwarnings = [
  'error',
]
markers = [
]

[build-system]
requires = ["poetry_core>=1.0.8"]
build-backend = "poetry.core.masonry.api"
